Fix the parsing of sequence IDs in user provided OTU tables
This was due to the fact that the "ID" of a sequence in a FASTA file is only the very first word that appears after the ">" and until the first space. Whatever follows is considered part of the "description" of the sequence.

However I was not doing the same thing for the parsing of the OTU table, where the whole text would be considered the ID. I fixed the problem and added a test that runs with the test suite.