Fix the no hits sequences unreported when using VSEARCH
The problem where VSEARCH doesn't report the sequences that got no hits. This is due to the fact that even with the "---blast6out" option turned on, it doesn't actually reproduce the BLAST output format. In BLAST, sequences that no have no hits are mentioned in the output file as having no hits. In VSEARCH sequences with no hits just disappear entirely from the search results. But this should be fixed now.